Move local source into `/src/`.
