+++ demo/contacts/contacts.js
@@ -1,5 +1,5 @@
-import Layout from '../../layouts/layout-1d.js';
-import {VirtualScroller} from '../../virtual-scroller.js';
+import Layout from '../../src/layouts/layout-1d.js';
+import {VirtualScroller} from '../../src/virtual-scroller.js';
 
 const scrollMethod = document.createElement('div').scrollIntoViewIfNeeded ?
     'scrollIntoViewIfNeeded' :
