+++ demo/fallback.html
@@ -51,7 +51,7 @@
   </virtual-scroller>
 
   <script type="module">
-    import '../virtual-scroller-element.js';
+    import '../src/virtual-scroller-element.js';
     import './contacts/contact-element.js';
     import { ContactDataSource } from './contacts/contact-data-source.js';
