+++ demo/header-footer.html
@@ -30,7 +30,7 @@
   <virtual-scroller></virtual-scroller>
 
   <script type="module">
-    import '../virtual-scroller-element.js';
+    import '../src/virtual-scroller-element.js';
     import './contacts/contact-element.js';
 
     const virtualScroller = document.querySelector('virtual-scroller');
