+++ demo/html-spec/html-spec-viewer.js
@@ -1,6 +1,6 @@
 import {HtmlSpec} from '../../node_modules/streaming-spec/HtmlSpec.js';
 import {iterateStream} from '../../node_modules/streaming-spec/iterateStream.js';
-import {ItemSource, VirtualScrollerElement} from '../../virtual-scroller-element.js';
+import {ItemSource, VirtualScrollerElement} from '../../src/virtual-scroller-element.js';
 
 class HTMLSpecSource extends ItemSource {
   static fromArray(items) {
