+++ demo/layout.html
@@ -86,7 +86,7 @@
   </main>
 
   <script type="module">
-    import '../virtual-scroller-element.js';
+    import '../src/virtual-scroller-element.js';
 
     const main = document.querySelector('main');
     const scroller = document.querySelector('#scroller');
