+++ demo/loadmore.html
@@ -32,9 +32,9 @@
   <virtual-scroller></virtual-scroller>
 
   <script type="module">
-    import '../virtual-scroller-element.js';
+    import '../src/virtual-scroller-element.js';
     import './contacts/contact-element.js';
-    import { ItemSource } from '../item-source.js';
+    import { ItemSource } from '../src/item-source.js';
     import { ContactDataSource } from './contacts/contact-data-source.js';
 
     const dataSource = new ContactDataSource();
