+++ demo/scrolling.html
@@ -60,7 +60,7 @@
   <virtual-scroller id="scroller"></virtual-scroller>
 
   <script type="module">
-    import '../virtual-scroller-element.js';
+    import '../src/virtual-scroller-element.js';
 
     const virtualScroller = document.getElementById('scroller');
