+++ demo/sorting.html
@@ -39,7 +39,7 @@
   </virtual-scroller>
 
   <script type="module">
-    import { ItemSource } from '../virtual-scroller-element.js';
+    import { ItemSource } from '../src/virtual-scroller-element.js';
     import './contacts/contact-element.js';
 
     const virtualScroller = document.querySelector('virtual-scroller');
