+++ smoke/infinite/index.html
@@ -20,7 +20,7 @@
   </style>
   <virtual-scroller id="virtualScroller"></virtual-scroller>
   <script type="module">
-    import { ItemSource } from '../../virtual-scroller-element.js';
+    import { ItemSource } from '../../src/virtual-scroller-element.js';
 
     // Render as many items as possible --> infinite scrolling!
     virtualScroller.itemSource = new ItemSource({
